#972: o.c.e.Threads - extend IterableEnvelope
